+++ CHANGELOG.md
@@ -19,7 +19,14 @@
   - add `readonly` combinator
   - add `never` type
 - **Breaking Changes**
-  - remove `maybe` combinator
+  - remove `maybe` combinator, can be defined in userland as
+    ```ts
+    export function maybe<RT extends t.Any>(type: RT, name?: string): t.UnionType<[RT, typeof t.null], t.TypeOf<RT> | null> {
+      return t.union([type, t.null], name)
+    }
+    ```
+- **Polish**
+  - export `pathReporterFailure` function from default reporters
 - **Bug Fix**
   - revert pruning excess properties
   - revert `intersection` combinator accepting only `InterfaceType`s